ml/requirements.txt
scikit-learn==1.7.1
pandas==2.3.2
numpy==2.3.2
pymysql==1.1.1
joblib==1.4.2
python-dotenv==1.0.1